component finished and in place of the old one
