+++ apps/catalog/catalog/src/app/marketplace/title/list/list.component.html
@@ -50,7 +50,7 @@
     <ng-template filter label="Country of Origin" [form]="searchForm.originCountries">
       <group-multiselect
         scope="territories"
-        label="Select territoties"
+        label="Select Territories"
         displayAll="World"
         filterPlaceholder="Specify territories to filter"
         [control]="searchForm.get('originCountries')"
