+++ apps/financiers/financiers/src/app/marketplace/title/list/list.component.html
@@ -43,16 +43,13 @@
     </ng-template>
 
     <ng-template filter label="Country of Origin" [form]="searchForm.originCountries">
-      <mat-form-field appearance="outline">
-        <mat-label>Select Territories</mat-label>
-        <mat-icon matPrefix svgIcon="world"></mat-icon>
-        <static-group
-          [formControl]="searchForm.get('originCountries')"
-          scope="territories"
-          displayAll="World"
-          placeholder="Specify territory to filter">
-        </static-group>
-      </mat-form-field>
+      <group-multiselect
+        scope="territories"
+        label="Select territoties"
+        displayAll="World"
+        filterPlaceholder="Specify territories to filter"
+        [control]="searchForm.get('originCountries')">
+      </group-multiselect>
     </ng-template>
 
     <ng-template filter label="Language & Version" [form]="searchForm.languages">
