+++ libs/contract/src/lib/avails/filter/filter.component.html
@@ -1,36 +1,25 @@
-<mat-form-field appearance="outline" *ngIf="getControl('territories') as territories" >
-  <mat-label>Select Territories</mat-label>
-  <mat-icon matPrefix svgIcon="world"></mat-icon>
-  <static-group
-    [disabled]="disabled"
-    [formControl]="territories"
+<ng-container *ngIf="getControl('territories') as territories">
+  <group-multiselect
     scope="territories"
+    label="Select Territoties"
     displayAll="World"
-    placeholder="Specify territory to filter"
+    filterPlaceholder="Specify territories to filter"
     required
-    test-id="territories">
-  </static-group>
-  <mat-error *ngIf="form.hasError('required', 'territories')">
-    This field is mandatory
-  </mat-error>
-</mat-form-field>
+    [control]="territories">
+  </group-multiselect>
+</ng-container>
 
-<mat-form-field appearance="outline" *ngIf="getControl('medias') as medias">
-  <mat-label>Select Rights</mat-label>
-  <mat-icon matPrefix svgIcon="play_arrow"></mat-icon>
-  <static-group
-    [formControl]="medias"
-    [disabled]="disabled"
+<ng-container *ngIf="getControl('medias') as medias">
+  <group-multiselect
     scope="medias"
+    label="Select Rights"
     displayAll="All Rights"
+    filterPlaceholder="Specify right(s) to filter"
+    selectIcon="play_arrow"
     required
-    placeholder="Specify right(s) to filter"
-    test-id="medias">
-  </static-group>
-  <mat-error *ngIf="form.hasError('required', 'medias')">
-    This field is mandatory
-  </mat-error>
-</mat-form-field>
+    [control]="medias">
+  </group-multiselect>
+</ng-container>
 
 <div class="duration" *ngIf="getControl('duration') as duration" [formGroup]="duration" >
   <mat-form-field appearance="outline">
