+++ libs/contract/src/lib/avails/filter/filter.component.scss
@@ -20,7 +20,7 @@
   .duration {
     flex: 1;
   }
-  mat-form-field {
+  mat-form-field, group-multiselect {
     flex: 1;
     margin-right: 12px;
   }
