+++ libs/contract/src/lib/contract/holdback/form/form.component.scss
@@ -24,10 +24,6 @@ bf-form-table {
   }
 }
 
-static-group {
-  width: 100%;
-}
-
 mat-divider {
   margin: 16px 0;
 }
