+++ libs/contract/src/lib/contract/holdback/list/list.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ListComponent } from './list.component';
 
-import { StaticGroupModule } from '@blockframes/ui/static-autocomplete/group/group.module';
 import { JoinPipeModule, MaxLengthModule, ToGroupLabelPipeModule, } from '@blockframes/utils/pipes';
 import { TableModule } from '@blockframes/ui/list/table/table.module';
 
@@ -13,7 +12,6 @@ import { TableModule } from '@blockframes/ui/list/table/table.module';
   imports: [
     CommonModule,
     TableModule,
-    StaticGroupModule,
     JoinPipeModule,
     MaxLengthModule,
     ToGroupLabelPipeModule,
