+++ libs/contract/src/lib/contract/holdback/selection-modal/selection-modal.component.ts
@@ -6,6 +6,7 @@ import { HoldbackForm } from '../form';
 import { DetailedTermsComponent } from '@blockframes/contract/term/components/detailed/detailed.component';
 import { FormTableComponent } from '@blockframes/ui/form/table/form-table.component';
 import { createModalData } from '@blockframes/ui/global-modal/global-modal.component';
+import { Validators } from '@angular/forms';
 
 @Component({
   selector: 'holdback-selection',
@@ -35,7 +36,7 @@ export class SelectionModalComponent implements OnInit {
   ) {}
 
   ngOnInit() {
-    this.form = FormList.factory(this.data.holdbacks, (holdback) => new HoldbackForm(holdback), []);
+    this.form = FormList.factory(this.data.holdbacks, (holdback) => new HoldbackForm(holdback), Validators.required);
   }
 
   close() {
