+++ libs/ui/src/lib/static-autocomplete/group/group.component.html
@@ -1,41 +1,56 @@
-﻿<ng-container *ngIf="groups$ | async as groups">
-  <ng-container *ngIf="value$ | async as value">
-    <mat-select multiple [formControl]="form" [disabled]="disabled"
-      (openedChange)="onOpen($event)" >
-      <header>
-        <mat-checkbox color="primary" [indeterminate]="(value | getMode:groups) === 'indeterminate'"
-          [checked]="(value | getMode:groups) === 'checked'" (change)="checkAll($event.checked)" test-id="checked-all">
+<mat-form-field appearance="outline" [attr.test-id]="scope">
+  <mat-label>{{ label }}</mat-label>
+  <mat-icon *ngIf="selectIcon !== 'none'" matPrefix [svgIcon]="selectIcon"></mat-icon>
+  <mat-select multiple [formControl]="control" (openedChange)="onOpen()" [required]="required" test-id="filter-select">
+    <mat-select-trigger>{{
+      control.value.length === items.length ? displayAll : (control.value | toGroupLabel: scope | join)
+    }}</mat-select-trigger>
+    <header>
+      <mat-checkbox
+        color="primary"
+        [indeterminate]="indeterminate.all"
+        [checked]="checked.all"
+        (change)="checkAll($event.checked)"
+        test-id="checked-all"
+      ></mat-checkbox>
+      <input
+        [placeholder]="filterPlaceholder"
+        [formControl]="search"
+        (keydown)="$event.stopPropagation()"
+        #searchInput
+        test-id="search-input-control"
+      />
+      <mat-icon *ngIf="!search.value" svgIcon="search"></mat-icon>
+      <button *ngIf="!!search.value" mat-icon-button type="button" (click)="resetSearch()" matTooltip="Clear text">
+        <mat-icon svgIcon="refresh"></mat-icon>
+      </button>
+    </header>
+    <ng-container *ngFor="let group of groups">
+      <div [ngClass]="{ hidden: !selectable[group.label] }" class="mat-option group">
+        <mat-checkbox
+          color="primary"
+          [indeterminate]="indeterminate[group.label]"
+          [checked]="checked[group.label]"
+          (change)="checkGroup(group.label, selectable, $event.checked)"
+          [attr.test-id]="group.label"
+          >{{ group.label }}
         </mat-checkbox>
-        <input (keydown)="cancelSpace($event)" #inputEl class="mat-input-element" [placeholder]="placeholder" [formControl]="search" test-id="search-input-control" />
-        <mat-icon *ngIf="!search.value" svgIcon="search"></mat-icon>
-        <button *ngIf="search.value" mat-icon-button type="button" (click)="search.reset()" matTooltip="Clear text">
-          <mat-icon svgIcon="refresh"></mat-icon>
+        <button mat-icon-button color="primary" (click)="toggleVisibility(group.label)">
+          <mat-icon [svgIcon]="visible[group.label] ? 'expand_more' : 'expand_less'"></mat-icon>
         </button>
-      </header>
-      <ng-container *ngFor="let group of filteredGroups$ | async; trackBy: trackByLabel">
-        <div matRipple (click)="checkGroup(group, !checkbox.checked, $event)"
-          class="subheader mat-option mat-focus-indicator mat-option-multiple mat-optgroup-label" [attr.test-id]="group.label">
-          <mat-checkbox #checkbox="matCheckbox" color="primary"
-            [indeterminate]="(value | getMode:group) === 'indeterminate'"
-            [checked]="(value | getMode:group) === 'checked'" (change)="checkGroup(group, $event.checked)"
-            class="mat-option-pseudo-checkbox">
-          </mat-checkbox>
-          <b class="mat-option-text">{{ group.label }}</b>
-          <button mat-icon-button (click)="hideGroup(group, $event)">
-            <mat-icon *ngIf="!hidden[group.label]" svgIcon="expand_less"></mat-icon>
-            <mat-icon *ngIf="hidden[group.label]" svgIcon="expand_more"></mat-icon>
-          </button>
-        </div>
         <mat-divider></mat-divider>
-        <ng-container *ngFor="let item of group.items">
-          <mat-option class="item" [ngClass]="{ hidden: hidden[group.label] }" [value]="item" [attr.test-id]="item">
-            {{ item | toLabel: scope }}
-          </mat-option>
-        </ng-container>
+      </div>
+      <ng-container *ngFor="let item of group.items">
+        <mat-option
+          [value]="item"
+          class="item"
+          [ngClass]="{ hidden: !selectable[item] || visible[group.label] }"
+          [attr.test-id]="item"
+        >
+          {{ item | toLabel: scope }}
+        </mat-option>
       </ng-container>
-      <mat-select-trigger>
-        {{ value | toGroupLabel: scope: displayAll|join }}
-      </mat-select-trigger>
-    </mat-select>
-  </ng-container>
-</ng-container>
+    </ng-container>
+  </mat-select>
+  <mat-error *ngIf="control.hasError('required')"> This field is mandatory </mat-error>
+</mat-form-field>
