+++ libs/ui/src/lib/static-autocomplete/group/group.module.ts
@@ -1,17 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { StaticGroupComponent, GetModePipe } from './group.component';
-import { ReactiveFormsModule } from "@angular/forms";
-import { MatFormFieldModule } from "@angular/material/form-field";
-import { MatSelectModule } from "@angular/material/select";
-import { MatButtonModule } from "@angular/material/button";
-import { MatInputModule } from "@angular/material/input";
-import { MatCheckboxModule } from "@angular/material/checkbox";
-import { MatIconModule } from "@angular/material/icon";
-import { MatDividerModule } from "@angular/material/divider";
-import { MatRippleModule } from "@angular/material/core";
-import { MatTooltipModule } from '@angular/material/tooltip';
+import { ReactiveFormsModule } from '@angular/forms';
 import { JoinPipeModule, ToGroupLabelPipeModule, ToLabelModule } from '@blockframes/utils/pipes';
+import { GroupMultiselectComponent } from './group.component';
+
+// Material
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatSelectModule } from '@angular/material/select';
+import { MatButtonModule } from '@angular/material/button';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { MatDividerModule } from '@angular/material/divider';
 
 @NgModule({
   imports: [
@@ -22,15 +22,13 @@ import { JoinPipeModule, ToGroupLabelPipeModule, ToLabelModule } from '@blockfra
     MatButtonModule,
     MatIconModule,
     MatCheckboxModule,
-    MatInputModule,
     MatDividerModule,
-    MatRippleModule,
     MatTooltipModule,
     ToLabelModule,
     ToGroupLabelPipeModule,
     JoinPipeModule,
   ],
-  declarations: [StaticGroupComponent, GetModePipe],
-  exports: [StaticGroupComponent]
+  declarations: [GroupMultiselectComponent],
+  exports: [GroupMultiselectComponent],
 })
-export class StaticGroupModule { }
+export class GroupMultiselectModule {}
