--- libs/ui/src/lib/static-autocomplete/groupB/group.component.html
@@ -1,55 +1,0 @@
-<mat-form-field appearance="outline">
-  <mat-label>{{ label }}</mat-label>
-  <mat-icon matPrefix [svgIcon]="selectIcon"></mat-icon>
-  <mat-select multiple [formControl]="control" (openedChange)="onOpen()" test-id="filter-select">
-    <mat-select-trigger>{{
-      control.value.length === items.length ? displayAll : (control.value | toGroupLabel: scope | join)
-    }}</mat-select-trigger>
-    <header>
-      <mat-checkbox
-        color="primary"
-        [indeterminate]="indeterminate.all"
-        [checked]="checked.all"
-        (change)="checkAll($event.checked)"
-        test-id="checked-all"
-      ></mat-checkbox>
-      <input
-        [placeholder]="filterPlaceholder"
-        [formControl]="search"
-        (keydown)="$event.stopPropagation()"
-        #searchInput
-        test-id="search-input-control"
-      />
-      <mat-icon *ngIf="!search.value" svgIcon="search"></mat-icon>
-      <button *ngIf="!!search.value" mat-icon-button type="button" (click)="resetSearch()" matTooltip="Clear text">
-        <mat-icon svgIcon="refresh"></mat-icon>
-      </button>
-    </header>
-    <ng-container *ngFor="let group of groups">
-      <div [ngClass]="{ hidden: !selectable[group.label] }" class="mat-option group">
-        <mat-checkbox
-          color="primary"
-          [indeterminate]="indeterminate[group.label]"
-          [checked]="checked[group.label]"
-          (change)="checkGroup(group.label, selectable, $event.checked)"
-          [attr.test-id]="group.label"
-          >{{ group.label }}
-        </mat-checkbox>
-        <button mat-icon-button color="primary" (click)="toggleVisibility(group.label)">
-          <mat-icon [svgIcon]="visible[group.label] ? 'expand_more' : 'expand_less'"></mat-icon>
-        </button>
-        <mat-divider></mat-divider>
-      </div>
-      <ng-container *ngFor="let item of group.items">
-        <mat-option
-          [value]="item"
-          class="item"
-          [ngClass]="{ hidden: !selectable[item] || visible[group.label] }"
-          [attr.test-id]="item"
-        >
-          {{ item | toLabel: scope }}
-        </mat-option>
-      </ng-container>
-    </ng-container>
-  </mat-select>
-</mat-form-field>
