--- libs/ui/src/lib/static-autocomplete/groupB/group.component.scss
@@ -1,48 +1,0 @@
-mat-form-field {
-  width: 100%;
-}
-
-header {
-  display: flex;
-  align-items: center;
-  padding: 8px;
-  border-bottom: solid 1px var(--foreground-divider);
-  position: sticky;
-  top: 0;
-  background-color: var(--background-card);
-  z-index: 1;
-
-  mat-checkbox {
-    padding: 8px;
-  }
-}
-
-input {
-  margin: 0 8px;
-  flex: 1;
-  background-color: transparent;
-  border: none;
-  font: inherit;
-
-  &:focus {
-    outline: none;
-  }
-}
-
-.hidden {
-  display: none;
-}
-
-.item {
-  padding: 0 24px;
-}
-
-mat-checkbox{
-  width: 100%;
-}
-
-button{
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-}
