+++ libs/user/src/lib/auth/forms/preferences/preferences.component.html
@@ -1,25 +1,23 @@
 <ng-container *ngIf="form">
   <form [formGroup]="form">
 
-    <mat-form-field appearance="outline">
-      <mat-label>Select Rights</mat-label>
-      <static-group
-        [formControl]="form.get('medias')"
-        scope="medias"
-        displayAll="All Rights"
-        placeholder="Select right(s)">
-      </static-group>
-    </mat-form-field>
-    
-    <mat-form-field appearance="outline">
-      <mat-label>Select Territories</mat-label>
-      <static-group
-        [formControl]="form.get('territories')"
-        scope="territories"
-        displayAll="World"
-        placeholder="Select Territories">
-      </static-group>
-    </mat-form-field>
+    <group-multiselect
+      scope="medias"
+      label="Select Rights"
+      displayAll="All Rights"
+      filterPlaceholder="Specify right(s) to filter"
+      selectIcon="none"
+      [control]="form.get('medias')">
+    </group-multiselect>
+
+    <group-multiselect
+      scope="territories"
+      label="Select Territories"
+      displayAll="World"
+      filterPlaceholder="Specify territories to filter"
+      selectIcon="none"
+      [control]="form.get('territories')">
+    </group-multiselect>
 
     <chips-autocomplete [form]="form.get('languages')" scope="languages" withoutValues="all">
       <mat-label>Languages</mat-label>
